Add files via upload
This file is meant to replace the existing file in the usr/src/asterisk install folder.  It is from Asterisk v11 but I don't think there's much difference in the more recent versions - but check anyways!    

This patched file will allow you to use rotary telephones (pulse dialling) with vertical feature codes normally commencing with "*" - for example, *69, *70, etc can be accessed by dialling 1169, 1170, etc.

In addition, you are now able to place an audio file in "/var/lib/asterisk/sounds" where after a certain amount time, it will be played back to the user.  In this case, it is for a recording which reminds the user to hangup and then after a short while can be used to play the ROH (receiver off-hook) tone.